image-processing/requirements.txt
Pillow~=10.0.1
opencv-python~=4.8.1.78
numpy~=1.26.1
pandas~=2.1.3
rasterio~=1.3.9
scikit-image~=0.22.0
scikit-learn~=1.3.2